requirements.txt
aiogram~=3.3.0
asyncio~=3.4.3
python-dotenv~=1.0.0
APScheduler~=3.10.4
pandas~=2.1.4
openpyxl==3.1.2
